http: making http upstreams pluggable (#11327)
Mainly moving code, but finishes up the series of pluggable upstream PRs.

Additional Description: This unhides the configurable extension point to the cluster for selecting a connection pool and creating an upstream, which can be used for custom business logic in upstream creation.

Risk Level: medium (router refactor)
Testing: with prior PRs
Docs Changes: inline with APIs
Release Notes: added
