+++ source/common/router/upstream_request.cc
@@ -517,126 +517,5 @@ void UpstreamRequest::enableDataFromDownstreamForFlowControl() {
   }
 }
 
-void HttpConnPool::newStream(GenericConnectionPoolCallbacks* callbacks) {
-  callbacks_ = callbacks;
-  // It's possible for a reset to happen inline within the newStream() call. In this case, we
-  // might get deleted inline as well. Only write the returned handle out if it is not nullptr to
-  // deal with this case.
-  Http::ConnectionPool::Cancellable* handle =
-      conn_pool_->newStream(*callbacks->upstreamRequest(), *this);
-  if (handle) {
-    conn_pool_stream_handle_ = handle;
-  }
-}
-
-void TcpConnPool::onPoolReady(Tcp::ConnectionPool::ConnectionDataPtr&& conn_data,
-                              Upstream::HostDescriptionConstSharedPtr host) {
-  upstream_handle_ = nullptr;
-  Network::Connection& latched_conn = conn_data->connection();
-  auto upstream =
-      std::make_unique<TcpUpstream>(callbacks_->upstreamRequest(), std::move(conn_data));
-  callbacks_->onPoolReady(std::move(upstream), host, latched_conn.localAddress(),
-                          latched_conn.streamInfo());
-}
-
-bool HttpConnPool::cancelAnyPendingRequest() {
-  if (conn_pool_stream_handle_) {
-    conn_pool_stream_handle_->cancel(Tcp::ConnectionPool::CancelPolicy::Default);
-    conn_pool_stream_handle_ = nullptr;
-    return true;
-  }
-  return false;
-}
-
-absl::optional<Http::Protocol> HttpConnPool::protocol() const { return conn_pool_->protocol(); }
-
-void HttpConnPool::onPoolFailure(ConnectionPool::PoolFailureReason reason,
-                                 absl::string_view transport_failure_reason,
-                                 Upstream::HostDescriptionConstSharedPtr host) {
-  callbacks_->onPoolFailure(reason, transport_failure_reason, host);
-}
-
-void HttpConnPool::onPoolReady(Http::RequestEncoder& request_encoder,
-                               Upstream::HostDescriptionConstSharedPtr host,
-                               const StreamInfo::StreamInfo& info) {
-  conn_pool_stream_handle_ = nullptr;
-  auto upstream = std::make_unique<HttpUpstream>(*callbacks_->upstreamRequest(), &request_encoder);
-  callbacks_->onPoolReady(std::move(upstream), host,
-                          request_encoder.getStream().connectionLocalAddress(), info);
-}
-
-TcpUpstream::TcpUpstream(UpstreamRequest* upstream_request,
-                         Tcp::ConnectionPool::ConnectionDataPtr&& upstream)
-    : upstream_request_(upstream_request), upstream_conn_data_(std::move(upstream)) {
-  upstream_conn_data_->connection().enableHalfClose(true);
-  upstream_conn_data_->addUpstreamCallbacks(*this);
-}
-
-void TcpUpstream::encodeData(Buffer::Instance& data, bool end_stream) {
-  upstream_conn_data_->connection().write(data, end_stream);
-}
-
-void TcpUpstream::encodeHeaders(const Http::RequestHeaderMap&, bool end_stream) {
-  // Headers should only happen once, so use this opportunity to add the proxy
-  // proto header, if configured.
-  ASSERT(upstream_request_->parent().routeEntry()->connectConfig().has_value());
-  Buffer::OwnedImpl data;
-  auto& connect_config = upstream_request_->parent().routeEntry()->connectConfig().value();
-  if (connect_config.has_proxy_protocol_config()) {
-    const Network::Connection& connection = *upstream_request_->parent().callbacks()->connection();
-    Extensions::Common::ProxyProtocol::generateProxyProtoHeader(
-        connect_config.proxy_protocol_config(), connection, data);
-  }
-
-  if (data.length() != 0 || end_stream) {
-    upstream_conn_data_->connection().write(data, end_stream);
-  }
-
-  // TcpUpstream::encodeHeaders is called after the UpstreamRequest is fully initialized. Also use
-  // this time to synthesize the 200 response headers downstream to complete the CONNECT handshake.
-  Http::ResponseHeaderMapPtr headers{
-      Http::createHeaderMap<Http::ResponseHeaderMapImpl>({{Http::Headers::get().Status, "200"}})};
-  upstream_request_->decodeHeaders(std::move(headers), false);
-}
-
-void TcpUpstream::encodeTrailers(const Http::RequestTrailerMap&) {
-  Buffer::OwnedImpl data;
-  upstream_conn_data_->connection().write(data, true);
-}
-
-void TcpUpstream::readDisable(bool disable) {
-  if (upstream_conn_data_->connection().state() != Network::Connection::State::Open) {
-    return;
-  }
-  upstream_conn_data_->connection().readDisable(disable);
-}
-
-void TcpUpstream::resetStream() {
-  upstream_request_ = nullptr;
-  upstream_conn_data_->connection().close(Network::ConnectionCloseType::NoFlush);
-}
-
-void TcpUpstream::onUpstreamData(Buffer::Instance& data, bool end_stream) {
-  upstream_request_->decodeData(data, end_stream);
-}
-
-void TcpUpstream::onEvent(Network::ConnectionEvent event) {
-  if (event != Network::ConnectionEvent::Connected && upstream_request_) {
-    upstream_request_->onResetStream(Http::StreamResetReason::ConnectionTermination, "");
-  }
-}
-
-void TcpUpstream::onAboveWriteBufferHighWatermark() {
-  if (upstream_request_) {
-    upstream_request_->disableDataFromDownstreamForFlowControl();
-  }
-}
-
-void TcpUpstream::onBelowWriteBufferLowWatermark() {
-  if (upstream_request_) {
-    upstream_request_->enableDataFromDownstreamForFlowControl();
-  }
-}
-
 } // namespace Router
 } // namespace Envoy
