+++ source/common/router/upstream_request.h
@@ -162,145 +162,5 @@ private:
   Event::TimerPtr max_stream_duration_timer_;
 };
 
-class HttpConnPool : public GenericConnPool, public Http::ConnectionPool::Callbacks {
-public:
-  // GenericConnPool
-  HttpConnPool(Upstream::ClusterManager& cm, const RouteEntry& route_entry,
-               absl::optional<Http::Protocol> downstream_protocol,
-               Upstream::LoadBalancerContext* ctx) {
-    conn_pool_ = cm.httpConnPoolForCluster(route_entry.clusterName(), route_entry.priority(),
-                                           downstream_protocol, ctx);
-  }
-  bool valid() const { return conn_pool_ != nullptr; }
-
-  void newStream(GenericConnectionPoolCallbacks* callbacks) override;
-  bool cancelAnyPendingRequest() override;
-  absl::optional<Http::Protocol> protocol() const override;
-  Upstream::HostDescriptionConstSharedPtr host() const override { return conn_pool_->host(); }
-
-  // Http::ConnectionPool::Callbacks
-  void onPoolFailure(ConnectionPool::PoolFailureReason reason,
-                     absl::string_view transport_failure_reason,
-                     Upstream::HostDescriptionConstSharedPtr host) override;
-  void onPoolReady(Http::RequestEncoder& callbacks_encoder,
-                   Upstream::HostDescriptionConstSharedPtr host,
-                   const StreamInfo::StreamInfo& info) override;
-
-private:
-  // Points to the actual connection pool to create streams from.
-  Http::ConnectionPool::Instance* conn_pool_{};
-  Http::ConnectionPool::Cancellable* conn_pool_stream_handle_{};
-  GenericConnectionPoolCallbacks* callbacks_{};
-};
-
-class TcpConnPool : public GenericConnPool, public Tcp::ConnectionPool::Callbacks {
-public:
-  TcpConnPool(Upstream::ClusterManager& cm, const RouteEntry& route_entry,
-              absl::optional<Http::Protocol>, Upstream::LoadBalancerContext* ctx) {
-    conn_pool_ = cm.tcpConnPoolForCluster(route_entry.clusterName(),
-                                          Upstream::ResourcePriority::Default, ctx);
-  }
-  bool valid() const { return conn_pool_ != nullptr; }
-  void newStream(GenericConnectionPoolCallbacks* callbacks) override {
-    callbacks_ = callbacks;
-    upstream_handle_ = conn_pool_->newConnection(*this);
-  }
-
-  bool cancelAnyPendingRequest() override {
-    if (upstream_handle_) {
-      upstream_handle_->cancel(ConnectionPool::CancelPolicy::Default);
-      upstream_handle_ = nullptr;
-      return true;
-    }
-    return false;
-  }
-  absl::optional<Http::Protocol> protocol() const override { return absl::nullopt; }
-  Upstream::HostDescriptionConstSharedPtr host() const override { return conn_pool_->host(); }
-  // Tcp::ConnectionPool::Callbacks
-  void onPoolFailure(ConnectionPool::PoolFailureReason reason,
-                     Upstream::HostDescriptionConstSharedPtr host) override {
-    upstream_handle_ = nullptr;
-    callbacks_->onPoolFailure(reason, "", host);
-  }
-
-  void onPoolReady(Tcp::ConnectionPool::ConnectionDataPtr&& conn_data,
-                   Upstream::HostDescriptionConstSharedPtr host) override;
-
-private:
-  Tcp::ConnectionPool::Instance* conn_pool_;
-  Tcp::ConnectionPool::Cancellable* upstream_handle_{};
-  GenericConnectionPoolCallbacks* callbacks_{};
-};
-
-class HttpUpstream : public GenericUpstream, public Http::StreamCallbacks {
-public:
-  HttpUpstream(UpstreamRequest& upstream_request, Http::RequestEncoder* encoder)
-      : upstream_request_(upstream_request), request_encoder_(encoder) {
-    request_encoder_->getStream().addCallbacks(*this);
-  }
-
-  // GenericUpstream
-  void encodeData(Buffer::Instance& data, bool end_stream) override {
-    request_encoder_->encodeData(data, end_stream);
-  }
-  void encodeMetadata(const Http::MetadataMapVector& metadata_map_vector) override {
-    request_encoder_->encodeMetadata(metadata_map_vector);
-  }
-  void encodeHeaders(const Http::RequestHeaderMap& headers, bool end_stream) override {
-    request_encoder_->encodeHeaders(headers, end_stream);
-  }
-  void encodeTrailers(const Http::RequestTrailerMap& trailers) override {
-    request_encoder_->encodeTrailers(trailers);
-  }
-
-  void readDisable(bool disable) override { request_encoder_->getStream().readDisable(disable); }
-
-  void resetStream() override {
-    request_encoder_->getStream().removeCallbacks(*this);
-    request_encoder_->getStream().resetStream(Http::StreamResetReason::LocalReset);
-  }
-
-  // Http::StreamCallbacks
-  void onResetStream(Http::StreamResetReason reason,
-                     absl::string_view transport_failure_reason) override {
-    upstream_request_.onResetStream(reason, transport_failure_reason);
-  }
-
-  void onAboveWriteBufferHighWatermark() override {
-    upstream_request_.disableDataFromDownstreamForFlowControl();
-  }
-
-  void onBelowWriteBufferLowWatermark() override {
-    upstream_request_.enableDataFromDownstreamForFlowControl();
-  }
-
-private:
-  UpstreamRequest& upstream_request_;
-  Http::RequestEncoder* request_encoder_{};
-};
-
-class TcpUpstream : public GenericUpstream, public Tcp::ConnectionPool::UpstreamCallbacks {
-public:
-  TcpUpstream(UpstreamRequest* upstream_request, Tcp::ConnectionPool::ConnectionDataPtr&& upstream);
-
-  // GenericUpstream
-  void encodeData(Buffer::Instance& data, bool end_stream) override;
-  void encodeMetadata(const Http::MetadataMapVector&) override {}
-  void encodeHeaders(const Http::RequestHeaderMap&, bool end_stream) override;
-  void encodeTrailers(const Http::RequestTrailerMap&) override;
-  void readDisable(bool disable) override;
-  void resetStream() override;
-
-  // Tcp::ConnectionPool::UpstreamCallbacks
-  void onUpstreamData(Buffer::Instance& data, bool end_stream) override;
-  void onEvent(Network::ConnectionEvent event) override;
-  void onAboveWriteBufferHighWatermark() override;
-  void onBelowWriteBufferLowWatermark() override;
-
-private:
-  UpstreamRequest* upstream_request_;
-  Tcp::ConnectionPool::ConnectionDataPtr upstream_conn_data_;
-};
-
 } // namespace Router
 } // namespace Envoy
