+++ source/extensions/upstreams/http/generic/config.cc
@@ -1,6 +1,7 @@
 #include "extensions/upstreams/http/generic/config.h"
 
-#include "common/router/upstream_request.h"
+#include "extensions/upstreams/http/http/upstream_request.h"
+#include "extensions/upstreams/http/tcp/upstream_request.h"
 
 namespace Envoy {
 namespace Extensions {
@@ -13,10 +14,12 @@ Router::GenericConnPoolPtr GenericGenericConnPoolFactory::createGenericConnPool(
     absl::optional<Envoy::Http::Protocol> downstream_protocol,
     Upstream::LoadBalancerContext* ctx) const {
   if (is_connect) {
-    auto ret = std::make_unique<Router::TcpConnPool>(cm, route_entry, downstream_protocol, ctx);
+    auto ret = std::make_unique<Upstreams::Http::Tcp::TcpConnPool>(cm, is_connect, route_entry,
+                                                                   downstream_protocol, ctx);
     return (ret->valid() ? std::move(ret) : nullptr);
   }
-  auto ret = std::make_unique<Router::HttpConnPool>(cm, route_entry, downstream_protocol, ctx);
+  auto ret = std::make_unique<Upstreams::Http::Http::HttpConnPool>(cm, is_connect, route_entry,
+                                                                   downstream_protocol, ctx);
   return (ret->valid() ? std::move(ret) : nullptr);
 }
