+++ source/extensions/upstreams/http/http/config.cc
@@ -1,6 +1,6 @@
 #include "extensions/upstreams/http/http/config.h"
 
-#include "common/router/upstream_request.h"
+#include "extensions/upstreams/http/http/upstream_request.h"
 
 namespace Envoy {
 namespace Extensions {
@@ -9,10 +9,10 @@ namespace Http {
 namespace Http {
 
 Router::GenericConnPoolPtr HttpGenericConnPoolFactory::createGenericConnPool(
-    Upstream::ClusterManager& cm, bool, const Router::RouteEntry& route_entry,
+    Upstream::ClusterManager& cm, bool is_connect, const Router::RouteEntry& route_entry,
     absl::optional<Envoy::Http::Protocol> downstream_protocol,
     Upstream::LoadBalancerContext* ctx) const {
-  auto ret = std::make_unique<Router::HttpConnPool>(cm, route_entry, downstream_protocol, ctx);
+  auto ret = std::make_unique<HttpConnPool>(cm, is_connect, route_entry, downstream_protocol, ctx);
   return (ret->valid() ? std::move(ret) : nullptr);
 }
