+++ test/common/router/BUILD
@@ -328,26 +328,6 @@ envoy_cc_test(
 )
 
 envoy_cc_test(
-    name = "upstream_request_test",
-    srcs = ["upstream_request_test.cc"],
-    deps = [
-        "//source/common/buffer:buffer_lib",
-        "//source/common/router:router_lib",
-        "//source/common/upstream:upstream_includes",
-        "//source/common/upstream:upstream_lib",
-        "//test/common/http:common_lib",
-        "//test/mocks/http:http_mocks",
-        "//test/mocks/network:network_mocks",
-        "//test/mocks/router:router_mocks",
-        "//test/mocks/server:server_mocks",
-        "//test/mocks/upstream:upstream_mocks",
-        "//test/test_common:environment_lib",
-        "//test/test_common:simulated_time_system_lib",
-        "//test/test_common:utility_lib",
-    ],
-)
-
-envoy_cc_test(
     name = "header_formatter_test",
     srcs = ["header_formatter_test.cc"],
     deps = [
